// src/app/model/implemetation.model.ts
export class ImplementationDetails {
    constructor() {}

    ptsRef?: string;

    awardedProjectValueStatus?: string;
    awardedProjectValueDate?: string;
    expenditureStatus?: string;
    expenditureDate?: string;
    appointedServiceProviderStatus?: string;
    appointedServiceProviderDate?: string;
    kickoffMeetingStatus?: string;
    kickoffMeetingDate?: string;
    purchaseOrderStatus?: string;
    purchasedOrderDate?: string;
    projectCharterStatus?: string;
    projectCharterDate?: string;
    serviceLevelStatus?: string;
    serviceLevelDate?: string;
    informationGatheringStatus?: string;
    informationGatheringDate?: string;
    architecturalDesignStatus?: string;
    architecturalDesignDate?: string;
    softwareDeliveredStatus?: string;
    softwareDeliveredDate?: string;
    hardwareDeliveredStatus?: string;
    hardwareDeliveredDate?: string;
    installationStatus?: string;
    installationDate?: string;
    systemDevelopmentStatus?: string;
    systemDevelopmentDate?: string;
    solutionTestingStatus?: string;
    solutionTestingDate?: string;
    progressStatus?: string;
    progressDate?: string;
    challengesStatus?: string;
    challengesDate?: string;

    remarks?: string;
    remarksEx?: string;
    remarksAppointed?: string;
    remarksKickoff?: string;
    remarksPurchase?: string;
    remarksCharter?: string;
    remarksSLA?: string;
    remarksURS?: string;
    remarksFRS?: string;
    remarksSoftware?: string;
    remarksHardware?: string;
    remarksInstallation?: string;
    remarksDevelopment?: string;
    remarksTesting?: string;
    remarksProgress?: string;
    remarksClose?: string;
}